Move addLayer and addFrame to map store
